Library: 스타일링 라이브러리 테스트 / Vanilla-Extract, Turbopack과 호환성 문제로 Styled-Components 사용 결정
