+++ apps/blog/app/components/Header.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-
+// TODO: Styled-Components 적용 결정
 
 const Header = () => {
   return (
