+++ apps/blog/app/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import Header from "./Header";
+import Header from "./components/Header";
 import ThreeJs from "./components/ThreeJs";
 
 export default function Home() {
